Plutarch.Internal.Numeric, collect PNum and PIntegral into it
