+++ Plutarch/Positive.hs
@@ -15,10 +15,11 @@ import Data.Text (pack)
 import GHC.Generics (Generic)
 import Plutarch.Builtin (PAsData, PData, PIsData, pdata)
 import Plutarch.Builtin.Bool (pif)
-import Plutarch.Integer (PInteger, PIntegral)
+import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Eq (PEq)
 import Plutarch.Internal.Lift (DeriveNewtypePLiftable, PLiftable, PLifted (PLifted))
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
+import Plutarch.Internal.Numeric (PIntegral, PNum (pfromInteger, (#-)))
 import Plutarch.Internal.Ord (POrd, PPartialOrd ((#<=)))
 import Plutarch.Internal.Other (pto)
 import Plutarch.Internal.PLam (plam)
@@ -39,7 +40,6 @@ import Plutarch.Internal.Term (
  )
 import Plutarch.Internal.TermCont (runTermCont, tcont)
 import Plutarch.Maybe (PMaybe (PJust, PNothing))
-import Plutarch.Num (PNum (pfromInteger, (#-)))
 import Plutarch.Show (PShow, pshow)
 import Plutarch.Trace (ptraceInfoError)
 import Plutarch.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'), ptryFrom)
