+++ Plutarch/Show.hs
@@ -38,10 +38,11 @@ import Plutarch.ByteString (
   psliceBS,
   punsafeIntegerToByte,
  )
-import Plutarch.Integer (PInteger, PIntegral (pquot, prem))
+import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
 import Plutarch.Internal.Lift (pconstant)
+import Plutarch.Internal.Numeric (PIntegral (pquot, prem))
 import Plutarch.Internal.Ord ((#<))
 import Plutarch.Internal.Other (
   pfix,
