+++ plutarch.cabal
@@ -95,6 +95,7 @@ library
     Plutarch.Internal.Generic
     Plutarch.Internal.Lift
     Plutarch.Internal.Newtype
+    Plutarch.Internal.Numeric
     Plutarch.Internal.Ord
     Plutarch.Internal.Other
     Plutarch.Internal.PLam
@@ -110,7 +111,6 @@ library
     Plutarch.List
     Plutarch.Maybe
     Plutarch.Monadic
-    Plutarch.Num
     Plutarch.Pair
     Plutarch.Positive
     Plutarch.Prelude
